+++ look-and-feel/css/src/Accordion/Accordion.scss
@@ -1,209 +1,74 @@
 @use "../common/common" as common;
 
-.af-accordion {
-  font-family: common.$font-family-base;
-  font-size: 1.125rem;
-  line-height: 1.4rem;
+.af-accordion-client {
+  display: block;
+  margin: 0;
+  padding: 0;
+  border-radius: 8px;
+  font-family: "Source Sans Pro", arial, sans-serif;
+  background-color: common.$color-white;
 
-  &__summary {
-    display: grid;
-    padding: 1rem;
-    grid-template-areas: "tag-date-container title-container . value arrow";
-    grid-template-columns: repeat(2, auto) 1fr repeat(2, auto);
-    align-items: center;
-    column-gap: 2.5rem;
-    font-weight: 600;
-    color: common.$color-gray-900;
-    cursor: pointer;
+  &__item {
+    &:not(:last-child) {
+      border-bottom: 1px solid #e9ecf2;
+    }
 
-    p {
+    & > h3 {
+      width: 100%;
       margin: 0;
-    }
-
-    .af-accordion {
-      &__icon {
-        display: grid;
-        grid-area: icon;
-        align-items: center;
-      }
-
-      &__title-container {
-        display: grid;
-        grid-area: title-container;
-        grid-template: "icon title" "icon subtitle";
-        gap: 0.125rem 1rem;
-
-        &:not(:has(.af-accordion__icon)) {
-          column-gap: 0;
-        }
-
-        &:not(:has(.af-accordion__subtitle)) {
-          row-gap: 0;
-        }
-
-        .af-accordion {
-          &__title {
-            grid-area: title;
-          }
-
-          &__subtitle {
-            grid-area: subtitle;
-            font-size: 1rem;
-            font-weight: 400;
-            line-height: 1.25rem;
-            color: common.$color-gray-700;
-          }
-        }
-      }
-
-      &__value {
-        display: grid;
-        min-width: 9.375rem;
-        grid-area: value;
-        justify-items: center;
-        color: common.$color-green-600;
-      }
-
-      &__tag-date-container {
-        display: grid;
-        padding-right: 3.93rem;
-        grid-area: tag-date-container;
-        grid-template: "tag" "date";
-        row-gap: 0.5rem;
-
-        .af-accordion {
-          &__tag-container {
-            justify-items: center;
-          }
-
-          &__date {
-            display: grid;
-            align-items: center;
-            color: common.$color-gray-700;
-          }
-        }
-      }
-    }
-
-    &::after {
-      display: inline-block;
-      width: 2rem;
-      height: 2rem;
-      margin-left: auto;
-      grid-area: arrow;
-      flex-shrink: 0;
-      background-image: url("data:image/svg+xml,%3Csvg%20width%3D%2232%22%20height%3D%2232%22%20viewBox%3D%220%200%2032%2032%22%20fill%3D%22none%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%3E%0A%3Cg%20clip-path%3D%22url(%23clip0_14583_27882)%22%3E%0A%3Cpath%20d%3D%22M24%2013.3331L22.12%2011.4531L16%2017.5598L9.88%2011.4531L8%2013.3331L16%2021.3331L24%2013.3331Z%22%20fill%3D%22%2300008F%22%2F%3E%0A%3C%2Fg%3E%0A%3Cdefs%3E%0A%3CclipPath%20id%3D%22clip0_14583_27882%22%3E%0A%3Crect%20width%3D%2232%22%20height%3D%2232%22%20fill%3D%22white%22%20transform%3D%22matrix(0%201%20-1%200%2032%200)%22%2F%3E%0A%3C%2FclipPath%3E%0A%3C%2Fdefs%3E%0A%3C%2Fsvg%3E%0A");
-      background-repeat: no-repeat;
-      background-size: 100%;
-      transition: transform 0.2s ease-in-out;
-      content: "";
-    }
-
-    &--title-first {
-      grid-template-areas: "title-container . tag-date-container value arrow";
-      grid-template-columns: repeat(1, auto) 1fr repeat(3, auto);
-
-      .af-accordion {
-        &__tag-date-container {
-          padding: 0;
-          grid-template: "tag date";
-          column-gap: inherit;
-
-          .af-accordion {
-            &__tag-container {
-              padding: 0 1.688rem;
-            }
-
-            &__date {
-              padding: 0 3.125rem;
-            }
-          }
-        }
-
-        &__value {
-          min-width: auto;
-        }
-      }
-    }
-
-    .af-accordion__tag-date-container:has(> :only-child) {
-      row-gap: 0;
-    }
-  }
-
-  &[open] > .af-accordion__summary::after {
-    transform: rotate(180deg);
-  }
-
-  &__separator {
-    margin: 0;
-    border: 0;
-    border-top: 1px solid common.$color-gray-300;
-  }
-
-  &__content {
-    padding: 1rem;
-    font-weight: 400;
-    color: common.$color-black;
-  }
-}
-
-@media (width <= #{common.$breakpoint-md}) {
-  .af-accordion__summary--title-first {
-    column-gap: 1.875rem;
-
-    .af-accordion__tag-date-container .af-accordion {
-      &__tag-container,
-      &__date {
-        padding: 0;
-      }
-    }
-  }
-}
-
-@media (width <= #{common.$breakpoint-sm}) {
-  .af-accordion {
-    &__summary {
-      grid-template-areas: "tag-date-container . arrow" "title-container . arrow" "value . arrow";
-      grid-template-columns: repeat(1, auto) 1fr repeat(1, auto);
-      row-gap: 0.25rem;
-      font-size: 1rem;
+      font-size: 1.125rem;
       font-weight: 600;
-      line-height: 1.25rem;
 
-      .af-accordion {
-        &__title-container {
-          justify-content: start;
-
-          .af-accordion {
-            &__subtitle {
-              font-size: 0.875rem;
-              line-height: 1.1rem;
-            }
-          }
-        }
-
-        &__tag-date-container {
-          padding-right: 0;
-          grid-template: "tag date";
-          gap: 0 0.5rem;
-        }
-
-        &__value {
-          justify-content: start;
-        }
+      &:has(button[aria-expanded="true"]) {
+        border-bottom: none;
+        color: common.$color-axa;
       }
 
-      &--title-first {
-        grid-template-areas: "icon title-container . arrow" "icon tag-date-container . arrow" "icon value . arrow";
-        grid-template-columns: repeat(2, auto) 1fr repeat(1, auto);
-        column-gap: 0.75rem;
+      & > button {
+        display: flex;
+        width: 100%;
+        padding: 15px 20px;
+        border: 0;
+        align-items: center;
+        font-size: 1.125rem;
+        font-weight: 600;
+        text-align: start;
+        color: inherit;
+        background-color: inherit;
+        cursor: pointer;
+
+        &:focus {
+          outline: unset;
+        }
+
+        &:focus-visible {
+          outline: 2px solid common.$color-blue-3;
+          outline-offset: 3px;
+        }
+
+        &::after {
+          width: 32px;
+          height: 32px;
+          margin-left: auto;
+          flex-shrink: 0;
+          background-image: url("data:image/svg+xml,%3Csvg%20width%3D%2232%22%20height%3D%2232%22%20viewBox%3D%220%200%2032%2032%22%20fill%3D%22none%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%3E%0A%3Cg%20clip-path%3D%22url(%23clip0_14583_27882)%22%3E%0A%3Cpath%20d%3D%22M24%2013.3331L22.12%2011.4531L16%2017.5598L9.88%2011.4531L8%2013.3331L16%2021.3331L24%2013.3331Z%22%20fill%3D%22%2300008F%22%2F%3E%0A%3C%2Fg%3E%0A%3Cdefs%3E%0A%3CclipPath%20id%3D%22clip0_14583_27882%22%3E%0A%3Crect%20width%3D%2232%22%20height%3D%2232%22%20fill%3D%22white%22%20transform%3D%22matrix(0%201%20-1%200%2032%200)%22%2F%3E%0A%3C%2FclipPath%3E%0A%3C%2Fdefs%3E%0A%3C%2Fsvg%3E%0A");
+          background-repeat: no-repeat;
+          background-size: 100%;
+          transition: transform 0.2s ease-in-out;
+          content: "";
+        }
+
+        &[aria-expanded="true"] {
+          &::after {
+            transform: rotate(180deg);
+          }
+        }
       }
     }
 
-    &__content {
-      font-size: 0.875rem;
-      line-height: 1.1rem;
+    & > [role="region"] {
+      margin-bottom: 1rem;
+      padding: 0 20px;
     }
   }
 }
