+++ look-and-feel/css/src/Card/Card.scss
@@ -37,7 +37,7 @@
     width: 100%;
   }
 
-  &--accordion {
+  &--new-accordion {
     padding: 0;
     box-shadow: none;
   }
