+++ look-and-feel/css/src/look-and-feel.scss
@@ -13,6 +13,7 @@
 @use "Modal/Modal";
 @use "Alert/Alert";
 @use "Link/Link";
+@use "NewAccordion/NewAccordion";
 @use "Pagination/Pagination";
 @use "Pagination/PaginationDesktop";
 @use "Pagination/PaginationMobile";
