+++ look-and-feel/react/src/Accordion/Accordion.mdx
@@ -5,69 +5,20 @@ import * as AccordionStories from "./Accordion.stories";
 
 # Accordion
 
-## Import
+## How to use
+
+To use the `Accordion` component import it like that:
 
 ```tsx
 import { Accordion } from "@axa-fr/design-system-look-and-feel-react";
-```
 
-## Uses
-
-You can force to open or close the accordion with the `isOpen` prop. (if no `onClick` prop is given the component will work like natively)
-
-And when the user clicks on the summary to open or close the accordion you can use the `onClick` function prop to do something at that moment. (you will get the native event as `onClick` parameter)
-
-You can change the `className` or add a `classModifier` if you want to custom the component.
-
-You can also add/change the props of the `Tag` component inside of the `Accordion`.
-
-### Accordion
-
-```tsx
-export const AccordionComponent = () => (
-  <Accordion
-    icon={<Svg src={callIcon} />}
-    title="Titre onglet"
-    subtitle="Titre onglet"
-    tag="En attente"
-    date="01/01/2021"
-    value="+ 686,00 €"
-  >
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam semper magna et
-    tempor blandit. Nulla vitae eros a odio pretium gravida. Sed eget tortor nec
-    massa lobortis bibendum. Morbi eget ligula porttitor, euismod odio
-    vestibulum, porta massa. Aenean vel venenatis tellus, sed iaculis nisl.
+const MyComponent = () => (
+  <Accordion>
+    <Accordion.Item title="Title 1">Content 1</Accordion.Item>
+    <Accordion.Item title="Title 2">Content 2</Accordion.Item>
+    <Accordion.Item title="Title 3">Content 3</Accordion.Item>
   </Accordion>
 );
 ```
 
 <Canvas of={AccordionStories.AccordionStory} />
-
-<Controls of={AccordionStories.AccordionStory} />
-
-### Accordion with Card
-
-```tsx
-export const AccordionWithCardComponent = () => (
-  <Card classModifier="large accordion" tabIndex={undefined}>
-    <Accordion
-      icon={<Svg src={callIcon} />}
-      title="Titre onglet"
-      subtitle="Titre onglet"
-      tag="En attente"
-      date="01/01/2021"
-      value="+ 686,00 €"
-    >
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam semper magna
-      et tempor blandit. Nulla vitae eros a odio pretium gravida. Sed eget
-      tortor nec massa lobortis bibendum. Morbi eget ligula porttitor, euismod
-      odio vestibulum, porta massa. Aenean vel venenatis tellus, sed iaculis
-      nisl.
-    </Accordion>
-  </Card>
-);
-```
-
-<Canvas of={AccordionStories.AccordionWithCardStory} />
-
-<Controls of={AccordionStories.AccordionWithCardStory} />
