+++ look-and-feel/react/src/Accordion/Accordion.stories.tsx
@@ -1,51 +1,40 @@
-import callIcon from "@material-symbols/svg-400/outlined/call-fill.svg";
 import { Meta, StoryObj } from "@storybook/react";
-import type { ComponentProps } from "react";
-import { Accordion } from ".";
-import { Card, Svg } from "../../client";
+import { Accordion } from "./Accordion";
 
 const meta: Meta<typeof Accordion> = {
   component: Accordion,
   title: "Components/Accordion",
+  parameters: {
+    layout: "fullscreen",
+  },
 };
 
 export default meta;
 
-const defaultArgs = {
-  icon: <Svg src={callIcon} />,
-  title: "Titre onglet",
-  subtitle: "Titre onglet",
-  tag: "En attente",
-  date: "01/01/2021",
-  value: "+ 686,00 €",
-  children: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam semper magna et tempor blandit. 
-              Nulla vitae eros a odio pretium gravida. Sed eget tortor nec massa lobortis bibendum. Morbi eget 
-              ligula porttitor, euismod odio vestibulum, porta massa. Aenean vel venenatis tellus, sed iaculis nisl.`,
-  isOpen: undefined,
-  onClick: undefined,
-  isTitleFirst: true,
+type StoryProps = {
+  numberOfItems: number;
 };
 
-export const AccordionStory: StoryObj<ComponentProps<typeof Accordion>> = {
+export const AccordionStory: StoryObj<StoryProps> = {
   name: "Accordion",
-  render: (args) => <Accordion {...args} />,
-  args: defaultArgs,
-  argTypes: {
-    date: { control: "text" },
+  render: ({ numberOfItems }) => {
+    return (
+      <Accordion>
+        {Array.from({ length: numberOfItems }, (_, i) => (
+          <Accordion.Item key={i} title={`Title ${i + 1}`}>
+            Content {i + 1}
+          </Accordion.Item>
+        ))}
+      </Accordion>
+    );
   },
-};
 
-export const AccordionWithCardStory: StoryObj<
-  ComponentProps<typeof Accordion>
-> = {
-  name: "Accordion With Card",
-  render: (args) => (
-    <Card classModifier="large accordion" tabIndex={undefined}>
-      <Accordion {...args} />
-    </Card>
-  ),
-  args: defaultArgs,
+  args: {
+    numberOfItems: 3,
+  },
   argTypes: {
-    date: { control: "text" },
+    numberOfItems: {
+      control: { type: "range", min: 1, max: 10 },
+    },
   },
 };
