--- look-and-feel/react/src/Accordion/index.ts
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-look-and-feel-css/dist/Accordion/Accordion.scss";
-
-export { Accordion } from "./Accordion";
