+++ look-and-feel/react/src/NewAccordion/NewAccordion.mdx
@@ -1,0 +1,73 @@
+import { Controls, Canvas, Meta } from "@storybook/blocks";
+import * as NewAccordionStories from "./NewAccordion.stories";
+
+<Meta of={NewAccordionStories} name="NewAccordion" />
+
+# NewAccordion
+
+## Import
+
+```tsx
+import { NewAccordion } from "@axa-fr/design-system-look-and-feel-react";
+```
+
+## Uses
+
+You can force to open or close the accordion with the `isOpen` prop. (if no `onClick` prop is given the component will work like natively)
+
+And when the user clicks on the summary to open or close the accordion you can use the `onClick` function prop to do something at that moment. (you will get the native event as `onClick` parameter)
+
+You can change the `className` or add a `classModifier` if you want to custom the component.
+
+You can also add/change the props of the `Tag` component inside of the `NewAccordion`.
+
+### NewAccordion
+
+```tsx
+export const NewAccordionComponent = () => (
+  <NewAccordion
+    icon={<Svg src={callIcon} />}
+    title="Titre onglet"
+    subtitle="Titre onglet"
+    tag="En attente"
+    date="01/01/2021"
+    value="+ 686,00 €"
+  >
+    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam semper magna et
+    tempor blandit. Nulla vitae eros a odio pretium gravida. Sed eget tortor nec
+    massa lobortis bibendum. Morbi eget ligula porttitor, euismod odio
+    vestibulum, porta massa. Aenean vel venenatis tellus, sed iaculis nisl.
+  </NewAccordion>
+);
+```
+
+<Canvas of={NewAccordionStories.NewAccordionStory} />
+
+<Controls of={NewAccordionStories.NewAccordionStory} />
+
+### NewAccordion with Card
+
+```tsx
+export const NewAccordionWithCardComponent = () => (
+  <Card classModifier="large new-accordion" tabIndex={undefined}>
+    <NewAccordion
+      icon={<Svg src={callIcon} />}
+      title="Titre onglet"
+      subtitle="Titre onglet"
+      tag="En attente"
+      date="01/01/2021"
+      value="+ 686,00 €"
+    >
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam semper magna
+      et tempor blandit. Nulla vitae eros a odio pretium gravida. Sed eget
+      tortor nec massa lobortis bibendum. Morbi eget ligula porttitor, euismod
+      odio vestibulum, porta massa. Aenean vel venenatis tellus, sed iaculis
+      nisl.
+    </NewAccordion>
+  </Card>
+);
+```
+
+<Canvas of={NewAccordionStories.NewAccordionWithCardStory} />
+
+<Controls of={NewAccordionStories.NewAccordionWithCardStory} />
