+++ look-and-feel/react/src/NewAccordion/index.ts
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-look-and-feel-css/dist/NewAccordion/NewAccordion.scss";
+
+export { NewAccordion } from "./NewAccordion";
